ocr/requirements.txt
boto3==1.40.69
fastapi==0.123.5
protobuf==6.33.1
pydantic==2.12.5
pytest==9.0.1
requests==2.32.5
uvicorn==0.38.0
google-genai==1.52.0